requirements.txt
cogdb>=3.6.4
openai
python-dotenv
